+++ apps/web/src/pages/login.tsx
@@ -1,11 +1,11 @@
-import clsx from "clsx";
 import Logo from "components/Logo";
 import { signIn } from "next-auth/react";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { useForm } from "react-hook-form";
-import { FaGithub } from "react-icons/fa";
+import { FaGithub, FaGoogle } from "react-icons/fa";
 import { MdOutlineEmail } from "react-icons/md";
+import { twMerge } from "tailwind-merge";
 
 const DEFAULT_CALLBACK_URL = "/projects";
 
@@ -22,14 +22,14 @@ export default function LoginPage() {
 
   const { register, handleSubmit, formState } = useForm<{ email: string }>();
   return (
-    <main className="flex min-h-screen items-center justify-center bg-pink-100">
-      <div className="max-w-2xl rounded-md bg-white p-8 text-gray-900 shadow-2xl">
+    <main className="flex min-h-screen items-center justify-center bg-primary-background text-primary-foreground">
+      <div className="max-w-2xl rounded-md border border-accent-background-muted bg-primary-background p-8 text-primary-foreground shadow-2xl shadow-accent-background">
         <Link href="/" className="mb-12 block text-center">
           <Logo as="h1" />
         </Link>
         <div className="flex flex-col">
           <button
-            className={clsx(
+            className={twMerge(
               baseSocialButtonStyles,
               "bg-[#323a4d] text-white hover:bg-[#4b5672]"
             )}
@@ -41,6 +41,20 @@ export default function LoginPage() {
           >
             <FaGithub className="mr-2" size={24} />
             Login with GitHub
+          </button>
+          <button
+            className={twMerge(
+              baseSocialButtonStyles,
+              "mt-2 bg-[rgba(66,_133,_244,_1)] text-white hover:bg-[rgba(66,_133,_244,_0.8)]"
+            )}
+            onClick={() =>
+              signIn("google", {
+                callbackUrl,
+              })
+            }
+          >
+            <FaGoogle className="mr-2" size={24} />
+            Login with Google
           </button>
           <p className="my-6 text-center">or</p>
           <form
@@ -55,8 +69,8 @@ export default function LoginPage() {
               {...register("email", {
                 required: "Please enter an email",
               })}
-              className={clsx(
-                "mb-3 w-full rounded-md border px-6 py-3 outline-none",
+              className={twMerge(
+                "mb-3 w-full rounded-md border bg-transparent px-6 py-3 text-primary-foreground outline-none focus:border-accent-background focus:ring-0",
                 formState.errors.email ? "border-red-500" : "border-gray-400"
               )}
               placeholder="Email Address"
@@ -70,9 +84,9 @@ export default function LoginPage() {
             )}
             <button
               type="submit"
-              className={clsx(
+              className={twMerge(
                 baseSocialButtonStyles,
-                "bg-gray-200 text-gray-900 hover:bg-gray-300"
+                "bg-accent-background text-accent-foreground hover:bg-accent-background-hover"
               )}
             >
               <MdOutlineEmail className="mr-2" size={24} />